shuffle around files
