+++ lsl-sys/Cargo.toml
@@ -3,6 +3,7 @@ name = "lsl-sys"
 version = "0.1.0"
 authors = ["Andrew Moffat <arwmoffat@gmail.com>"]
 edition = "2018"
+links = "lsl"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
